Refactor and addition of new protocol feature: processing instructions.
Removed the client SocketPool in favor of storing connection state in a
greenlet-local on the client object itself. This makes it easier to
reason about whether we've closed a connection, while preserving the
greenlet-local behavior of the pool.

Added a new abstraction "processing instructions" which can precede an
arbitrary command, and can be used to somehow instruct the server to do
something. The initial use-case is to specify the db index for a one-off
command invocation.

Also added "max_dbs" as a command-line option and added logic to strip
comments from json config files.